checkpoint: refactor client connection into channel
